Restore blobs using first blob backup at or before a specified date.
Otherwise fail.  The repozo script does the same.

We used to pick the first blob backup *after* the specified date,
because we assumed that the user would specify the exact date that is
in the filestorage backup.
Note that the timestamp of the filestorage and blobstorage backups may be
a few seconds apart.
In the new situation, the user should pick the date of the blob backup
or slightly later.