Automatically remove old blobs backups that have no corresponding filestorage backup.
We compare the timestamp of the oldest filestorage backup with the timestamps of the
blob backups.  This can be the name, if you use ``blob_timestamps = true``,
or the modification date of the blob backup.

This means that the ``keep_blob_days`` option is ignored, unless you use ``only_blobs = true``.